Refactoring and documentation
