+++ src/test/java/io/github/thebusybiscuit/slimefun4/utils/biomes/TestBiomeMapParser.java
@@ -1,8 +1,5 @@
 package io.github.thebusybiscuit.slimefun4.utils.biomes;
 
-import java.util.function.Function;
-
-import javax.annotation.Nonnull;
 import javax.annotation.ParametersAreNonnullByDefault;
 
 import org.bukkit.NamespacedKey;
@@ -22,8 +19,8 @@ import be.seeseemelk.mockbukkit.MockBukkit;
 
 class TestBiomeMapParser {
 
-    private static final Function<JsonElement, String> AS_STRING = JsonElement::getAsString;
-    private static final Function<JsonElement, Integer> AS_INT = JsonElement::getAsInt;
+    private static final BiomeDataConverter<String> AS_STRING = JsonElement::getAsString;
+    private static final BiomeDataConverter<Integer> AS_INT = JsonElement::getAsInt;
 
     private static Slimefun plugin;
     private static NamespacedKey key;
@@ -96,68 +93,8 @@ class TestBiomeMapParser {
         assertMisconfiguration(AS_INT, "[{\"value\": 1, \"biomes\": [\"" + biome.getKey() + "\"]}, {\"value\": 2, \"biomes\": [\"" + biome.getKey() + "\"]}]");
     }
 
-    @Test
-    @DisplayName("Test different value converters")
-    void testValueConversion() {
-        Biome biome = Biome.OCEAN;
-
-        String stringValue = "hello world";
-        BiomeMap<String> stringMap = createBiomeMap(AS_STRING, "[{\"value\": \"" + stringValue + "\", \"biomes\": [\"" + biome.getKey() + "\"]}]");
-        Assertions.assertEquals(stringValue, stringMap.get(biome));
-
-        int intValue = 1024;
-        BiomeMap<Integer> intMap = createBiomeMap(AS_INT, "[{\"value\": \"" + intValue + "\", \"biomes\": [\"" + biome.getKey() + "\"]}]");
-        Assertions.assertEquals(intValue, intMap.get(biome));
-    }
-
-    @Test
-    @DisplayName("Test working BiomeMap from JSON")
-    void testBiomeMapFromJson() {
-        Biome biome1 = Biome.JUNGLE;
-        Biome biome2 = Biome.OCEAN;
-        Biome biome3 = Biome.DESERT;
-        BiomeMap<Integer> biomes = createBiomeMap(AS_INT, "[{\"value\":1,\"biomes\":[\"" + biome1.getKey() + "\", \"" + biome2.getKey() + "\"]}, {\"value\":2, \"biomes\":[\"" + biome3.getKey() + "\"]}]");
-
-        Assertions.assertTrue(biomes.contains(biome1));
-        Assertions.assertEquals(1, biomes.get(biome1));
-
-        Assertions.assertTrue(biomes.contains(biome2));
-        Assertions.assertEquals(1, biomes.get(biome2));
-
-        Assertions.assertTrue(biomes.contains(biome3));
-        Assertions.assertEquals(2, biomes.get(biome3));
-    }
-
-    @Test
-    @DisplayName("Test working BiomeMap manually")
-    void testManualBiomeMap() {
-        BiomeMap<String> biomes = new BiomeMap<>(key);
-        Biome biome = Biome.OCEAN;
-        String value = "Under the sea";
-
-        Assertions.assertTrue(biomes.put(biome, value));
-        Assertions.assertTrue(biomes.contains(biome));
-        Assertions.assertEquals(value, biomes.get(biome));
-
-        Assertions.assertTrue(biomes.remove(biome));
-        Assertions.assertFalse(biomes.contains(biome));
-    }
-
     @ParametersAreNonnullByDefault
-    private @Nonnull <T> BiomeMap<T> createBiomeMap(Function<JsonElement, T> function, String json) {
-        BiomeMapParser<T> parser = new BiomeMapParser<>(key, function);
-
-        try {
-            parser.read(json);
-        } catch (BiomeMapException e) {
-            Assertions.fail(e);
-        }
-
-        return parser.buildBiomeMap();
-    }
-
-    @ParametersAreNonnullByDefault
-    private <T> void assertMisconfiguration(Function<JsonElement, T> function, String json) {
+    private <T> void assertMisconfiguration(BiomeDataConverter<T> function, String json) {
         BiomeMapParser<T> parser = new BiomeMapParser<>(key, function);
         Assertions.assertThrows(BiomeMapException.class, () -> parser.read(json));
     }
